requirements.txt
google-cloud-pubsub==1.1.0
google-api-core==1.16.0
grpcio==1.26.0